### 2018-07-03 #234 @GuangchuangYu
CRAN package - scholar
The package provides functions to extract citation data and co-author network from Google Scholar.

```
Package: scholar
    person("Guangchuang", "Yu",   email = "guangchuangyu@gmail.com",   role = c("aut", "cre"), comment = c(ORCID = "0000-0002-6485-8781")),
    person("James", "Keirstead",  email = "james.keirstead@gmail.com", role = "aut"),
    person("Gregory", "Jefferis", email = "jefferis@gmail.com",        role = "ctb"),
    person("Jorge", "Cimentada", email = "cimentadaj@gmail.com",        role = "ctb"),
    person("Max", "Czapanskiy", email = "maxczapanskiy@gmail.com",        role = "ctb")
    )
Maintainer: Guangchuang Yu <guangchuangyu@gmail.com>
Version: 0.1.7
Title: Analyse Citation Data from Google Scholar
Description: Provides functions to extract citation data from Google
    Scholar.  Convenience functions are also provided for comparing
    multiple scholars and predicting future h-index values.
    R.cache,
    httr,
    rvest,
    xml2,
    tidygraph,
    ggraph,
    ggplot2
    prettydoc,
    roxygen2

```

<https://github.com/jkeirstead/scholar>

[e.g., "data extraction, because the package parses a scientific data file format"]

data retrieval, because the package extract data from online service.

Users who want to analyze citation data.

No (to my knowledge).

- [ x contains a vignette with examples of its essential functions and uses.

- [x] Do you intend for this package to go on CRAN?  (already on CRAN)
