### 2018-07-03 #234 @GuangchuangYu
@jkeirstead is the original author that created the package. James's work is out of academia and have no time to further develop/maintain the package.

I am the new maintainer of the package. However, I should be admitted that I am not the expert of web technologies.

As this package was contributed by several developers, including @cimentadaj @DominiqueMakowski @abfleishman @dfalster and @ jefferis, it would be great to move the package to ROpenSci and maintained by the community.
