### 2018-07-03 #234 @maelle
Thanks for your submission @guangchuangyu! Just to clarify, do you intend to keep maintaining the package if it is onboarded, or are you looking for a new maintainer?
