### 2018-07-03 #234 @maelle
Well for onboarding it'd be great if you remained maintainer, and you could look for a new maintainer later if/when you realize you don't have time any more and there's a good candidate. Does that sound good? Otherwise the maintainer change would have to happen before reviews, not during the onboarding process. 😸
