### 2018-07-03 #234 @maelle
Also note that the package will need some unit tests and continuous integration before the submission process starts. More details [here](https://ropensci.github.io/dev_guide/building.html). We can put this submission on hold while you add them. 😺
