### 2018-07-03 #234 @GuangchuangYu
thanks @maelle.

Sounds good and will add unit tests soon.
