### 2018-07-03 #234 @maelle
@jefferis :wave:, @sckott can probably tell you more, since he maintains `vcr` and `webmockr` and has a [WIP book about http testing with `crul`](https://github.com/ropensci/http-testing-book). We don't have guidance about API package testing in our gitbook yet, but there was [a thread about it in our forum](https://discuss.ropensci.org/t/best-practices-for-testing-api-packages/460).
