### 2018-07-03 #234 @jefferis
Thanks @maelle. My conclusion from looking at those is that we either:
* switch to crul and use vcr
* roll our own
* use https://cran.r-project.org/web/packages/httptest/index.html
@sckott any additional comments?
