### 2018-07-03 #234 @sckott
switching to `crul` is the easiest option if you want mocking and/or caching of real requests/responses. `crul` has some other benefits in that it has a pagination helper in case you need to do that, and has async built in if you need to do that. I've got a number of packages on CRAN already using `vcr` - see the reverse suggests here https://cran.rstudio.com/web/packages/vcr/

I am working on adapters in `webmockr` and `vcr` for both `curl` and `httr`, but they aren't quite ready yet.

I've not tried httptest or other caching options, so can't comment on those
