### 2018-07-19 dd43745
get_journalrank
