+++ R/publications.r
@@ -212,6 +212,11 @@ get_oldest_article <- function(id) {
 ##' @export
 ##' @author Dominique Makowski and Guangchuang Yu
 get_impactfactor <- function(journals, max.distance = 0.05) {
+    get_journal_stats(journals, max.distance, impactfactor)
+}
+
+
+get_journal_stats <- function(journals, max.distance, source_data, col = "Journal") {
     journals <- as.character(journals)
     index <- rep(NA, length(journals))
 
@@ -222,7 +227,7 @@ get_impactfactor <- function(journals, max.distance = 0.05) {
         }
 
         closest <- agrep(journal,
-                         impactfactor$Journal,
+                         source_data[[col]],
                          max.distance = max.distance,
                          value = FALSE,
                          ignore.case = TRUE)
@@ -234,7 +239,7 @@ get_impactfactor <- function(journals, max.distance = 0.05) {
             ## index[i] <- closest[1]
 
 
-            j <- grep(paste0("^", journal, "$"), impactfactor$Journal[closest], ignore.case=TRUE)
+            j <- grep(paste0("^", journal, "$"), source_data[[col]][closest], ignore.case=TRUE)
             if (length(j) > 0) {
                 j <- j[1]
                 index[i] <- closest[j]
@@ -254,7 +259,7 @@ get_impactfactor <- function(journals, max.distance = 0.05) {
                           paste0("^", journal),
                           paste0(journal, "$"))
             for (pp in patterns) {
-                j <- get_hit(pp, impactfactor$Journal[closest])
+                j <- get_hit(pp, source_data[[col]][closest])
                 if (!is.null(j)) {
                     hit <- j
                     break
@@ -266,31 +271,19 @@ get_impactfactor <- function(journals, max.distance = 0.05) {
 
     }
 
-    return(impactfactor[index, ])
+    return(source_data[index, ])
 }
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-##' Get journal metrics.
+##' Get journal ranking.
 ##'
-##' Get journal metrics (impact factor) for a journal list.
+##' Get journal ranking for a journal list.
 ##'
 ##' @examples
 ##' library(scholar)
 ##'
 ##' id <- get_publications("bg0BZ-QAAAAJ&hl")
-##' impact <- get_journalmetrics(journals=id$journal)
+##' impact <- get_journalrank(journals=id$journal)
 ##'
 ##' id <- cbind(id, impact)
 ##'
@@ -299,68 +292,13 @@ get_impactfactor <- function(journals, max.distance = 0.05) {
 ##' Expressed either as integer, or as a fraction of the pattern length times the maximal transformation cost
 ##' (will be replaced by the smallest integer not less than the corresponding fraction), or a list with possible components
 ##'
-##' @return Journal metrics data.
+##' @return Journal ranking data.
 ##'
 ##' @import dplyr
 ##' @export
 ##' @author Dominique Makowski and Guangchuang Yu
-get_journalmetrics <- function(journals, max.distance = 0.05) {
-
-  journals <- as.character(journals)
-  index <- rep(NA, length(journals))
-  
-  for(i in seq_along(journals)) {
-    journal <- journals[i]
-    if(journal == ""){
-      next
-    }
-    
-    closest <- agrep(journal,
-                     journalrankings$Journal,
-                     max.distance = max.distance,
-                     value = FALSE,
-                     ignore.case = TRUE)
-    
-    if(!is.null(closest)){
-      ## agrep() returns all "close" matches
-      ## but unfortunately does not return the degree of closeness.
-      
-      ## index[i] <- closest[1]
-      
-      
[…]
