+++ NAMESPACE
@@ -7,6 +7,7 @@ export(get_citation_history)
 export(get_coauthors)
 export(get_complete_authors)
 export(get_impactfactor)
+export(get_journalrank)
 export(get_num_articles)
 export(get_num_distinct_journals)
 export(get_num_top_journals)
