### 2018-08-01 49e9c25
added author_order function
Attempts at finding author's order in authors list
