+++ R/scholar.r
@@ -10,16 +10,16 @@ utils::globalVariables(c("name"))
 ##' affiliation, their homepage (if specified), and a summary of their
 ##' key citation and impact metrics.  The scholar ID can be found by
 ##' searching Google Scholar at \url{http://scholar.google.com}.
-##' 
+##'
 ##' @param id 	a character string specifying the Google Scholar ID.
 ##' If multiple ids are specified, only the first value is used and a
 ##' warning is generated.  See the example below for how to profile
 ##' multiple scholars.
-##' 
+##'
 ##' @return 	a list containing the scholar's name, affiliation,
 ##' citations, impact metrics, fields of study, homepage and
 ##' the author's list of coauthors provided by Google Scholar.
-##' 
+##'
 ##' @examples {
 ##'    ## Gets profiles of some famous physicists
 ##'    ids <- c("xJaxiEEAAAAJ", "qj74uXkAAAAJ")
@@ -43,19 +43,19 @@ get_profile <- function(id) {
   ## but the number of rows seems to vary by OS
   stats <- tables[[1]]
   rows <- nrow(stats)
-  
+
   ## The personal info is in
   name <- page %>% html_nodes(xpath="//*/div[@id='gsc_prf_in']") %>% html_text()
   bio_info <- page %>% html_nodes(xpath="//*/div[@class='gsc_prf_il']") %>% html_text()
   affiliation <- bio_info[1]
- 
+
   ## Specialities (trim out HTML non-breaking space)
   specs <- iconv(bio_info[2], from="UTF8", to="ASCII")
   specs <- str_trim(tolower(str_split(specs, ",")[[1]]))
 
   ## Extract the homepage
-  homepage <- page %>% html_nodes(xpath="//*/div[@id='gsc_prf_ivh']//a/@href") %>% html_text() 
-  
+  homepage <- page %>% html_nodes(xpath="//*/div[@id='gsc_prf_ivh']//a/@href") %>% html_text()
+
   ## Grab all coauthors
   coauthors <- list_coauthors(id, n_coauthors = 20) # maximum availabe in profile
 
@@ -97,22 +97,22 @@ get_citation_history <- function(id) {
     vals <- page %>% html_nodes(xpath="//*/span[@class='gsc_g_al']") %>%
         html_text() %>% as.numeric()
     if(length(years)>length(vals)){
-      # Some years don't have citations. 
+      # Some years don't have citations.
       # We need to match the citation counts and years
       # <a href="javascript:void(0)" class="gsc_g_a" style="left:8px;height:5px;z-index:9">\n  <span class="gsc_g_al">2</span>\n</a>
-      style_tags=page %>% html_nodes(css = '.gsc_g_a') %>% 
+      style_tags=page %>% html_nodes(css = '.gsc_g_a') %>%
         html_attr('style')
       # these z indices seem to be the indices starting with the last year
       zindices=as.integer(stringr::str_match(style_tags, 'z-index:([0-9]+)')[,2])
       # empty vector of 0s
       allvals=integer(length=length(years))
-      # fill in 
+      # fill in
       allvals[zindices]=vals
       # and then reverse
       vals=rev(allvals)
     }
     df <- data.frame(year=years, cites=vals)
-  
+
     return(df)
 }
 
@@ -126,10 +126,10 @@ get_citation_history <- function(id) {
 ##' containing publication where applicable.  So a \emph{journal} here
 ##' might actually be a journal, a book, a report, or some other
 ##' publication outlet.
-##' 
+##'
 ##' @param id 	a character string giving the Google Scholar id
 ##' @return the number of distinct journals
-##' @export 
+##' @export
 get_num_distinct_journals <- function(id) {
   id <- tidy_id(id)
[…]
