### 2018-08-01 fca983a
fix
fixed the percentage computation
