### 2018-08-01 3faa001
rename it to author_position
I think this is more appropriate
