### 2018-08-20 9dceff5
Fixed issue where pagesize > 100 in get_publications would only return first page of results.
