### 2018-08-21 a4a5901
minor improvements
