### 2018-08-21 b8316f4
Added warning if pagesize > 100 in get_publications (thank you @jefferis).
