### 2018-08-22 0f5bd60
Revert "Minor Changes"
This reverts commit b0dbeba98753d151131b225e3737d6702178456e.
