### 2018-11-13 c5c7476
get_publications function: Added cstop and sortby arguments to allow users to get info on latest papers. To get info on latest paper set cstat = 0, cstop = 0 and sortby = "year".
