### 2019-07-03 0acd4d2
Add function to search for scholar ID by name and affiliation
