### 2019-07-06 15368d8
warn if problems finding scholar id unique matches
