### 2019-07-06 1927cda
Use get_resp not RCurl::getURL
* this handles the user token to avoid being frozen out
* also RCurl is no longer a dependency
