### 2019-07-06 8700b37
Switch order of name params / default args
* you might want to search by last name and affiliation (i.e. without a first name)
