### 2019-07-06 b8c308b
make most get_scholar_id egs donttest
* check starts to take a while
