### 2019-07-08 7c021e3
re-roxygenise
* missing from previous commit
* also update to latest release version of roxygen
* and state encoding
