+++ DESCRIPTION
@@ -34,4 +34,5 @@ Suggests:
     prettydoc,
     roxygen2
 VignetteBuilder: knitr
-RoxygenNote: 6.1.0
+RoxygenNote: 6.1.1
+Encoding: UTF-8
