+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(author_position)
 export(compare_scholar_careers)
 export(compare_scholars)
 export(get_article_cite_history)
