### 2019-07-08 f532527
Fix httr warning
* need to extract content explicitly from response
