+++ NAMESPACE
@@ -28,6 +28,7 @@ importFrom(dplyr,mutate)
 importFrom(dplyr,row_number)
 importFrom(dplyr,summarize)
 importFrom(httr,GET)
+importFrom(httr,content)
 importFrom(rvest,html_attr)
 importFrom(rvest,html_nodes)
 importFrom(rvest,html_table)
