+++ NAMESPACE
@@ -38,4 +38,5 @@ importFrom(stringr,str_replace)
 importFrom(stringr,str_split)
 importFrom(stringr,str_sub)
 importFrom(stringr,str_trim)
+importFrom(utils,tail)
 importFrom(xml2,read_html)
