+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(scholar)
+
+test_check("scholar")
