+++ DESCRIPTION
@@ -32,7 +32,8 @@ Imports:
 Suggests:
     knitr,
     prettydoc,
-    roxygen2
+    roxygen2,
+    testthat
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
 Encoding: UTF-8
