+++ DESCRIPTION
@@ -33,7 +33,7 @@ Suggests:
     knitr,
     prettydoc,
     roxygen2,
-    testthat
+    testthat (>= 2.1.0)
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
 Encoding: UTF-8
