### 2019-07-08 0e5f2d7
factor out scholar_handle function
