### 2019-07-08 b62711f
make get_resp use scholar_handle directly
