### 2019-07-08 94c587b
We shouldn't need to call scholar_handle in tidy_id
* get_resp will look after this
