### 2019-07-08 f01eed2
remove superfluous tidy_id calls
* get_publications already calls tidy_id internally
