### 2019-07-08 2487e0a
move utility functions into their own source file
