### 2019-07-08 2db188a
move coauthor functions into their own source file
* note some of them were in zzz initially
