+++ R/coauthors.R
@@ -1,0 +1,154 @@
+#' Gets the network of coauthors of a scholar
+#'
+##' @param id a character string specifying the Google Scholar ID.
+##' If multiple ids are specified, only the first value is used and a
+##' warning is generated.
+#' @param n_coauthors Number of coauthors to explore. This number should usually be between 1 and 10 as
+#' choosing many coauthors can make the network graph too messy.
+#' @param n_deep The number of degrees that you want to go down the network. When \code{n_deep} is equal to \code{1}
+#' then \code{grab_coauthor} will only grab the coauthors of Joe and Mary, so Joe -- > Mary --> All coauthors. This can get
+#' out of control very quickly if \code{n_deep} is set to \code{2} or above. The preferred number is \code{1}, the default.
+#'
+#' @details Considering that scraping each publication for all coauthors is error prone, \code{get_coauthors}
+#' grabs only the coauthors listed on the google scholar profile (on the bottom right of the profile),
+#' not from all publications.
+#'
+#' @return A data frame with two columns showing all authors and coauthors.
+#'
+#' @seealso \code{\link{plot_coauthors}}
+#' @export
+#'
+#' @examples
+#'
+#' \dontrun{
+#'
+#' library(scholar)
+#' coauthor_network <- get_coauthors('amYIKXQAAAAJ&hl')
+#' plot_coauthors(coauthor_network)
+#' }
+#'
+#'
+get_coauthors <- function(id, n_coauthors = 5, n_deep = 1) {
+    stopifnot(is.numeric(n_deep), length(n_deep) >= 1, n_deep != 0)
+    stopifnot(is.numeric(n_coauthors), length(n_coauthors) >= 1, n_coauthors != 0)
+    
+    all_coauthors <- list_coauthors(id, n_coauthors)
+    
+    empty_network <- replicate(n_deep, list())
+    
+    # Here I grab the id of the coauthors url because list_coauthors
+    # needs to accept ids and not urls
+    for (i in seq_len(n_deep)) {
+        if (i == 1)  {
+            empty_network[[i]] <- clean_network(grab_id(all_coauthors$coauthors_url),
+                                                n_coauthors)
+        } else {
+            empty_network[[i]] <- clean_network(grab_id(empty_network[[i - 1]]$coauthors_url),
+                                                n_coauthors)
+        }
+    }
+    
+    final_network <- rbind(all_coauthors, Reduce(rbind, empty_network))
+    final_network$author <- stringr::str_to_title(final_network$author)
+    final_network$coauthors <- stringr::str_to_title(final_network$coauthors)
+    final_network[c("author", "coauthors")]
+}
+
+
+#' Plot a network of coauthors
+#'
+#' @param network A data frame given by \code{\link{get_coauthors}}
+#' @param size_labels Size of the label names
+#'
+#' @return a \code{ggplot2} object but prints a plot as a side effect.
+#' @export
+#' @importFrom dplyr "%>%"
+#'
+#' @seealso \code{\link{get_coauthors}}
+#'
+#' @examples
+#' \dontrun{
+#' library(scholar)
+#' coauthor_network <- get_coauthors('amYIKXQAAAAJ&hl')
+#' plot_coauthors(coauthor_network)
+#' }
+plot_coauthors <- function(network, size_labels = 5) {
+    graph <- tidygraph::as_tbl_graph(network) %>%
+        mutate(closeness = suppressWarnings(tidygraph::centrality_closeness())) %>%
+        filter(name != "")
+    # to delete authors who have **no coauthors**, that is ""
+    
+    ggraph::ggraph(graph, layout = 'kk') +
[…]
