### 2019-07-08 a571135
export and rename get_resp -> get_scholar_resp
* better name if we are going to export it
