+++ R/scholar.r
@@ -36,7 +36,7 @@ get_profile <- function(id) {
   url <- compose_url(id, url_template)
 
   ## Generate a list of all the tables identified by the scholar ID
-  page <- get_resp(url) %>% read_html()
+  page <- get_scholar_resp(url) %>% read_html()
   tables <- page %>% html_table()
 
   ## The citation stats are in tables[[1]]$tables$stats
@@ -91,7 +91,7 @@ get_citation_history <- function(id) {
     url <- compose_url(id, url_template)
 
     ## A better way would actually be to read out the plot of citations
-    page <- get_resp(url) %>% read_html()
+    page <- get_scholar_resp(url) %>% read_html()
     years <- page %>% html_nodes(xpath="//*/span[@class='gsc_g_t']") %>%
         html_text() %>% as.numeric()
     vals <- page %>% html_nodes(xpath="//*/span[@class='gsc_g_al']") %>%
@@ -270,7 +270,7 @@ get_scholar_id <- function(last_name="", first_name="", affiliation = NA) {
     last_name,
     '&hl=en&oi=ao'
   )
-  aa <- content(get_resp(url), as='text')
+  aa <- content(get_scholar_resp(url), as='text')
   ids <-
     stringr::str_extract_all(string = aa, pattern = ";user=[[:alnum:]]+[[:punct:]]")
   
