+++ R/getCompleteAuthors.R
@@ -15,7 +15,7 @@ get_complete_authors = function(id, pubid)
   url_template = "http://scholar.google.com/citations?view_op=view_citation&citation_for_view=%s:%s"
   url = sprintf(url_template, id, pubid)
   
-  url1<- get_resp(url) %>%
+  url1<- get_scholar_resp(url) %>%
     read_html
   auths=as.character(rvest::html_node(url1,".gsc_vcd_value") %>% rvest::html_text())
   return(auths)
