### 2019-07-08 aced74c
move tidy_id to utils source file
