+++ R/scholar-package.r
@@ -17,22 +17,3 @@
 ##' @name scholar
 ##' @docType package
 NULL
-
-
-##' Ensures that specified IDs are correctly formatted
-##'
-##' @param id a character string specifying the Google Scholar ID.
-##' If multiple ids are specified, only the first value is used and a
-##' warning is generated.
-##' @export
-##' @importFrom httr GET
-##' @keywords internal
-tidy_id <- function(id) {
-  if (length(id)!=1) {
-    id <- id[1]
-    msg <- sprintf("Only one ID at a time; retrieving %s", id)
-    warning(msg)
-  }
-
-  return(id)
-}
