### 2019-07-08 69e5c61
use roxygen generated package docs
