+++ R/scholar-package.r
@@ -1,19 +1,26 @@
-##' scholar
-##'
-##' The \code{scholar} package provides functions to extract citation
-##' data from Google Scholar.  There are also convenience functions
-##' for comparing multiple scholars and predicting h-index scores
-##' based on past publication records.
-##'
-##' @note A complementary set of Google Scholar functions can be found
-##' at  \url{http://biostat.jhsph.edu/~jleek/code/googleCite.r}.  The
-##' \code{scholar} package was developed independently.
-##'
-##' @source The package reads data from
-##' \url{http://scholar.google.com}.  Dates and citation counts are
-##' estimated and are determined automatically by a computer program.
-##' Use at your own risk.
-##' 
-##' @name scholar
-##' @docType package
+#' scholar
+#'
+#' The \code{scholar} package provides functions to extract citation
+#' data from Google Scholar.  There are also convenience functions
+#' for comparing multiple scholars and predicting h-index scores
+#' based on past publication records.
+#' 
+#' @section Data 
+#'
+#' @note A complementary set of Google Scholar functions can be found
+#' at  \url{http://biostat.jhsph.edu/~jleek/code/googleCite.r}.  The
+#' \code{scholar} package was developed independently.
+#'
+#' @source The package reads data from
+#' \url{http://scholar.google.com}.  Dates and citation counts are
+#' estimated and are determined automatically by a computer program.
+#' Use at your own risk.
+#' 
+#' @keywords internal
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
 NULL
