+++ man/scholar-package.Rd
@@ -1,7 +1,7 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/scholar-package.r
 \docType{package}
-\name{scholar}
+\name{scholar-package}
 \alias{scholar}
 \alias{scholar-package}
 \title{scholar}
@@ -22,3 +22,29 @@ A complementary set of Google Scholar functions can be found
 at  \url{http://biostat.jhsph.edu/~jleek/code/googleCite.r}.  The
 \code{scholar} package was developed independently.
 }
+\seealso{
+Useful links:
+\itemize{
+  \item Report bugs at \url{https://github.com/jkeirstead/scholar/issues}
+}
+
+}
+\author{
+\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (0000-0002-6485-8781)
+
+Authors:
+\itemize{
+  \item James Keirstead \email{james.keirstead@gmail.com}
+}
+
+Other contributors:
+\itemize{
+  \item Gregory Jefferis \email{jefferis@gmail.com} [contributor]
+  \item Gordon Getzinger [contributor]
+  \item Jorge Cimentada \email{cimentadaj@gmail.com} [contributor]
+  \item Max Czapanskiy \email{maxczapanskiy@gmail.com} [contributor]
+  \item Dominique Makowski \email{dom.makowski@gmail.com} [contributor]
+}
+
+}
+\keyword{internal}
