### 2019-07-08 a5a42c8
final get_scholar_resp rename fixes
