### 2019-07-08 5c2cb26
Mention Scholar cookies in package docs
